requirements.txt
numpy == 1.21.6
pandas == 1.2.3
pillow == 9.4.0
scikit-learn == 0.22.1
torchaudio == 0.13.1
torchvision == 0.14.1
tqdm == 4.65.0
transformers == 4.30.2

